update to eigen3
Conflicts:

	icp/CMakeLists.txt
	icp/icp.cpp
	icp/icp.hpp

+++ src/operators/MLSSimpleTraversability.cpp
@@ -358,9 +358,9 @@ void MLSSimpleTraversability::closeNarrowPassages(MLSSimpleTraversability::Outpu
         }
     }
 
-    for (int y = 0; y < map.getHeight(); ++y)
+    for (size_t y = 0; y < map.getHeight(); ++y)
     {
-        for (int x = 0; x < map.getWidth(); ++x)
+        for (size_t x = 0; x < map.getWidth(); ++x)
         {
             if (data[y][x] == 255)
                 data[y][x] = CLASS_OBSTACLE;
